+++ tests/test_utils.py
@@ -1,7 +1,9 @@
 import sys
 
 import pytest
+from collections.abc import Iterator
 from toy_crypto import utils
+from toy_crypto.types import Byte
 
 
 class TestUtils:
@@ -34,6 +36,15 @@ class TestUtils:
             bits = [bit for bit in utils.lsb_to_msb(n)]
             assert bits == expected
 
+    def test_hamming(self) -> None:
+        s1 = b"this is a test"
+        s2 = b"wokka wokka!!!"
+
+        hd = utils.hamming_distance(s1, s2)
+        assert hd == 37
+
+
+class TestXor:
     def test_xor(self) -> None:
         vectors = [
             (b"dusk", b"dawn", bytes.fromhex("00 14 04 05")),
@@ -53,32 +64,24 @@ class TestUtils:
             r = utils.xor(x, y)
             assert r == pad
 
-    def test_ixor(self) -> None:
-        vectors = [
-            (b"dusk", b"dawn", bytes.fromhex("00 14 04 05")),
-            (
-                b"Attack at dawn!",
-                bytes(10) + bytes.fromhex("00 14 04 05 00"),
-                b"Attack at dusk!",
-            ),
-            (
-                bytes(15),
-                bytes.fromhex("00 01 02"),
-                bytes.fromhex("00 01 02") * 5,
-            ),
-        ]
+    def test_iter(self) -> None:
+        pad = bytes.fromhex("00 36 5C")
+        p_modulus = len(pad)
+        single = list(range(0, 256))
+        s_modulus = len(single)
+        message: Iterator[Byte] = iter(single * 10)
 
-        for x, y, pad in vectors:
-            x = bytearray(x)
-            utils.ixor(x, y)
-            assert x == pad
+        iter_xor = utils.Xor(message, pad)
 
-    def test_hamming(self) -> None:
-        s1 = b"this is a test"
-        s2 = b"wokka wokka!!!"
+        m_idx = 0
+        p_idx = 0
+        for b in iter_xor:
+            expected = (m_idx % s_modulus) ^ pad[p_idx]
+            assert b == expected
+            m_idx += 1
+            p_idx = (p_idx + 1) % p_modulus
 
-        hd = utils.hamming_distance(s1, s2)
-        assert hd == 37
+        assert m_idx == s_modulus * 10
 
 
 if __name__ == "__main__":